Remove comprehensive README file and add .gitignore for project dependencies and environment files
